+++ apps/backoffice/src/config/index.ts
@@ -20,12 +20,10 @@ export type Configuration = {
 export function getConfiguration() {
   return {
     // IO Services CMS API configuration
-    API_SERVICES_CMS_URL: process.env
-      .NEXT_PUBLIC_API_SERVICES_CMS_URL as string,
+    API_SERVICES_CMS_URL: process.env.API_SERVICES_CMS_URL as string,
     API_SERVICES_CMS_BASE_PATH: process.env
-      .NEXT_PUBLIC_API_SERVICES_CMS_BASE_PATH as string,
-    API_SERVICES_CMS_MOCKING:
-      process.env.NEXT_PUBLIC_API_SERVICES_CMS_MOCKING === "true",
+      .API_SERVICES_CMS_BASE_PATH as string,
+    API_SERVICES_CMS_MOCKING: process.env.API_SERVICES_CMS_MOCKING === "true",
 
     // BackOffice backend for frontend API configuration
     API_BACKEND_BASE_URL: process.env