+++ apps/backoffice/src/pages/keys/index.tsx
@@ -1,20 +1,55 @@
+import { ApiKeys } from "@/components/api-keys";
 import { PageHeader } from "@/components/headers";
+import { Cidr } from "@/generated/api/Cidr";
+import { ManageKeyCIDRs } from "@/generated/api/ManageKeyCIDRs";
+import { SubscriptionKeyTypeEnum } from "@/generated/api/SubscriptionKeyType";
+import { SubscriptionKeys } from "@/generated/api/SubscriptionKeys";
+import useFetch from "@/hooks/use-fetch";
 import { AppLayout, PageLayout } from "@/layouts";
 import { useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { ReactElement } from "react";
+import { ReactElement, useEffect } from "react";
 
 const pageTitleLocaleKey = "routes.keys.title";
 const pageDescriptionLocaleKey = "routes.keys.description";
 
 export default function Keys() {
   const { t } = useTranslation();
+  const { data: mkData, fetchData: mkFetchData } = useFetch<SubscriptionKeys>();
+  const { data: acData, fetchData: acFetchData } = useFetch<ManageKeyCIDRs>();
+
+  const handleRotateKey = (keyType: SubscriptionKeyTypeEnum) => {
+    mkFetchData("regenerateManageKey", { keyType }, SubscriptionKeys);
+  };
+
+  const handleUpdateCidrs = (cidrs: string[]) => {
+    acFetchData(
+      "updateManageKeysAuthorizedCidrs",
+      { body: { cidrs: Array.from(cidrs || []).filter(Cidr.is) } },
+      ManageKeyCIDRs
+    );
+  };
+
+  useEffect(() => {
+    mkFetchData("getManageKeys", {}, SubscriptionKeys);
+    acFetchData("getManageKeysAuthorizedCidrs", {}, ManageKeyCIDRs);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
     <>
       <PageHeader
         title={pageTitleLocaleKey}
         description={pageDescriptionLocaleKey}
+      />
+      <ApiKeys
+        title={t("routes.keys.manage.title")}
+        description={t("routes.keys.manage.description")}
+        keys={mkData}
+        onRotateKey={handleRotateKey}
+        showAuthorizedCidrs
+        cidrs={(acData?.cidrs as unknown) as string[]}
+        onUpdateCidrs={handleUpdateCidrs}
       />
     </>
   );
@@ -36,3 +71,5 @@ Keys.getLayout = function getLayout(page: ReactElement) {
     </AppLayout>
   );
 };
+
+Keys.requiredPermissions = ["apiservicewrite"];