+++ apps/backoffice/src/pages/mocks/selfcare.tsx
@@ -52,7 +52,7 @@ export default function Selfcare() {
 
 export async function getStaticProps({ locale }: any) {
   // ! return a not found error in prod environemnt
-  if (process.env.NODE_ENV === "production") {
+  if (process.env.APP_ENV === "production") {
     return { notFound: true };
   }
   return {