+++ apps/io-services-cms-webapp/CHANGELOG.md
@@ -1,5 +1,29 @@
 # io-services-cms-webapp
 
+## 1.21.0
+
+### Minor Changes
+
+- 94eec1c: Moving Apim Client in the new External-Clients package
+
+### Patch Changes
+
+- 1572c54: Update openapi-codegen-ts to version ^13.1.0 in io-services-cms-webapp
+- Updated dependencies [94eec1c]
+  - @io-services-cms/external-clients@1.0.0
+
+## 1.20.9
+
+### Patch Changes
+
+- 9dfd38e: Fix Azure Deploy action
+
+## 1.20.8
+
+### Patch Changes
+
+- e47f151: Handle Jira Issue status name case change
+
 ## 1.20.7
 
 ### Patch Changes